Crate state_api

Structs

MessageLookup

Functions

state_account_key
returns the public key address of the given ID address

state_all_miner_faults
returns all non-expired Faults that occur within lookback epochs of the given tipset

state_call
runs the given message and returns its result without any persisted changes.

state_for_ts
returns a state tree given a tipset

state_get_actor
returns the indicated actor's nonce and balance.

state_get_network_name

state_get_receipt
returns the message receipt for the given message

state_lookup_id
retrieves the ID address of the given address

state_market_balance
looks up the Escrow and Locked balances of the given address in the Storage Market

state_miner_deadlines
returns all the proving deadlines for the given miner

state_miner_faults
returns a single non-expired Faults that occur within lookback epochs of the given tipset

state_miner_info
StateMinerInfo returns info about the indicated miner

state_miner_proving_deadline
calculates the deadline at some epoch for a proving period and returns the deadline-related calculations.

state_miner_proving_set
returns info about those sectors that a given miner is actively proving.

state_miner_recoveries
returns a bitfield indicating the recovering sectors of the given miner

state_miner_sectors
returns info about the given miner's sectors. If the filter bitfield is nil, all sectors are included. If the filterOut boolean is set to true, any sectors in the filter are excluded. If false, only those sectors in the filter are included.

state_pledge_collateral

state_reply
returns the result of executing the indicated message, assuming it was executed in the indicated tipset.

state_sector_info
returns the on-chain info for the specified miner's sector

state_sector_precommit_info
returns the PreCommit info for the specified miner's sector

state_wait_msg
looks back in the chain for a message. If not found, it blocks until the message arrives on chain, and gets to the indicated confidence depth.